01_postgres-mcp/track-b/mcp_server/requirements.txt
fastmcp>=2.13.3,<2.14
psycopg2-binary>=2.9.0
